refactor(api): move course controller to src
